first steps of b1
