+++ src/main/java/com/volomak/movieland/dao/jdbc/CoutryDaoImpl.java
@@ -2,6 +2,8 @@ package com.volomak.movieland.dao.jdbc;
 
 import com.volomak.movieland.dao.CountryDao;
 import com.volomak.movieland.entity.Country;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.PreparedStatementSetter;
@@ -15,6 +17,8 @@ import java.util.List;
 
 @Service
 public class CoutryDaoImpl implements CountryDao {
+    private final Logger log = LoggerFactory.getLogger(getClass());
+
     @Autowired
     private String getCountriesByMovieIdSQL;
 
@@ -28,6 +32,8 @@ public class CoutryDaoImpl implements CountryDao {
 
     @Override
     public List<Country> getByMovieId(int id) {
+        log.info("Start query to get countris with movie id {} from DB", id);
+        long startTime = System.currentTimeMillis();
         List<Country> countries =  jdbcTemplate.query(getCountriesByMovieIdSQL,
                 new PreparedStatementSetter() {
                     @Override
@@ -45,6 +51,7 @@ public class CoutryDaoImpl implements CountryDao {
                     }
                 }
         );
+        log.info("Finish query to get countries with movie id {} from DB. It took {} ms", id, System.currentTimeMillis() - startTime);
         return countries;
     }
