+++ src/main/java/com/volomak/movieland/dao/jdbc/ReviewDaoImpl.java
@@ -2,6 +2,8 @@ package com.volomak.movieland.dao.jdbc;
 
 import com.volomak.movieland.dao.ReviewDao;
 import com.volomak.movieland.entity.Review;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.PreparedStatementSetter;
@@ -15,6 +17,8 @@ import java.util.List;
 
 @Service
 public class ReviewDaoImpl implements ReviewDao {
+    private final Logger log = LoggerFactory.getLogger(getClass());
+
     @Autowired
     private String getReviewsByMovieIdSQL;
 
@@ -28,6 +32,8 @@ public class ReviewDaoImpl implements ReviewDao {
 
     @Override
     public List<Review> getByMovieId(int id) {
+        log.info("Start query to get reviews with movie id {} from DB", id);
+        long startTime = System.currentTimeMillis();
         List<Review> reviews = jdbcTemplate.query(getReviewsByMovieIdSQL,
                 new PreparedStatementSetter() {
                     @Override
@@ -45,6 +51,7 @@ public class ReviewDaoImpl implements ReviewDao {
                 }
 
         );
+        log.info("Finish query to get reviews with movie id {} from DB. It took {} ms", id, System.currentTimeMillis() - startTime);
         return reviews;
     }
