--- src/main/java/com/volomak/movieland/dao/jdbc/mapper/GenreRowMapper.java
@@ -1,18 +1,0 @@
-package com.volomak.movieland.dao.jdbc.mapper;
-
-import com.volomak.movieland.entity.Genre;
-import org.springframework.jdbc.core.RowMapper;
-
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-public class GenreRowMapper implements RowMapper<Genre> {
-    @Override
-    public Genre mapRow(ResultSet resultSet, int i) throws SQLException {
-        Genre genre = new Genre();
-        genre.setId(resultSet.getInt("id"));
-        genre.setName(resultSet.getString("name_c"));
-        return genre;
-    }
-}
